refactored to be able to add generateObjectPage
