+++ packages/fe-fpm-writer/src/page/types.ts
@@ -1,6 +1,51 @@
 import type { CustomElement, InternalCustomElement } from '../common/types';
 
 /**
+ * Incoming navigation configuration.
+ */
+export interface Navigation {
+    /**
+     * Source of the navigation to the custom page
+     */
+    sourcePage: string;
+
+    /**
+     * Source entity that is use as the base part of the route
+     */
+    sourceEntity: string;
+
+    /**
+     * Name of the navigation entity
+     */
+    navEntity: string;
+
+    /**
+     * If set to true, then the generated route will point to a specific entity, otherwise, it will point to a list of the entities.
+     */
+    navKey?: boolean;
+}
+
+/**
+ * Configuration options for adding an object page.
+ */
+export interface ObjectPage {
+    /**
+     * Optional name of the object page. If not provided then the name will be generated based on the entity name.
+     */
+    name?: string;
+
+    /**
+     * Name of the entity used for the custom page.
+     */
+    entity: string;
+
+    /**
+     * Optional incoming navigation configuration. If provided then a navigation configuration is added to an existing page navigating to this custom page.
+     */
+    navigation?: Navigation;
+}
+
+/**
  * Configuration options for adding a custom page.
  */
 export interface CustomPage extends CustomElement {
@@ -12,22 +57,7 @@ export interface CustomPage extends CustomElement {
     /**
      * Optional incoming navigation configuration. If provided then a navigation configuration is added to an existing page navigating to this custom page.
      */
-    navigation?: {
-        /**
-         * Source of the navigation to the custom page
-         */
-        sourcePage: string;
-
-        /**
-         * Source entity that is use as the base part of the route
-         */
-        sourceEntity: string;
-
-        /**
-         * Name of the navigation entity
-         */
-        navEntity: string;
-
+    navigation?: Navigation & {
         /**
          * If set to true, then the generated route will point to a specific entity, otherwise, it will point to a list of the entities.
          */
@@ -45,19 +75,22 @@ export interface CustomPage extends CustomElement {
     };
 }
 
+export interface FCL {
+    /**
+     * The FCL flag is calculated to true if the existing app is configured to use the flexible column layout
+     */
+    fcl: boolean;
+
+    /**
+     * Represents the name of the control aggregation to be used for the custom page.
+     * This is only used if FCL is enabled and is calculated based on the optional sourcepage.
+     */
+    controlAggregation?: 'beginColumnPages' | 'midColumnPages' | 'endColumnPages';
+}
+
 /**
  * Additional internal configuration options that will be calculated based on the provided input as well as the target application.
  */
-export type InternalCustomPage = CustomPage &
-    InternalCustomElement & {
-        /**
-         * The FLC flag is calculated to true if the existing app is configured to use the flexible column layout
-         */
-        fcl: boolean;
+export type InternalCustomPage = CustomPage & InternalCustomElement & FCL;
 
-        /**
-         * Represents the name of the control aggregation to be used for the custom page.
-         * This is only used if FCL is enabled and is calculated based on the optional sourcepage.
-         */
-        controlAggregation?: 'beginColumnPages' | 'midColumnPages' | 'endColumnPages';
-    };
+export type InternalObjectPage = ObjectPage & FCL;
