+++ packages/fe-fpm-writer/test/unit/page.test.ts
@@ -3,7 +3,7 @@ import { create, Editor } from 'mem-fs-editor';
 import { join } from 'path';
 import { ManifestNamespace } from '@sap-ux/ui5-config';
 import { generateCustomPage, validateBasePath, CustomPage } from '../../src';
-import { validateCustomPageConfig } from '../../src/page';
+import { validatePageConfig } from '../../src/page';
 import { Manifest } from '../../src/common/types';
 import { FCL_ROUTER } from '../../src/common/defaults';
 
@@ -76,7 +76,7 @@ describe('CustomPage', () => {
             const manifest = JSON.parse(testAppManifest) as Manifest;
 
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).not.toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).not.toThrowError();
         });
 
         test('provided navigation config is not valid for existing manifest', () => {
@@ -86,29 +86,29 @@ describe('CustomPage', () => {
 
             manifest['sap.ui5']!.routing!.routes = [];
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
 
             delete manifest['sap.ui5']!.routing!.routes;
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
 
             manifest = JSON.parse(testAppManifest) as Manifest;
 
             delete manifest['sap.ui5']!.routing!.targets!['TestObjectPage'];
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
 
             delete manifest['sap.ui5']!.routing!.targets;
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
 
             delete manifest['sap.ui5']!.routing;
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
 
             delete manifest['sap.ui5'];
             fs.writeJSON(join(target, 'webapp/manifest.json'), manifest);
-            expect(() => validateCustomPageConfig(target, config, fs)).toThrowError();
+            expect(() => validatePageConfig(target, config, fs)).toThrowError();
         });
     });
